+++ Microsoft.Azure.Cosmos/src/Query/QueryMetrics.cs
@@ -157,7 +157,7 @@ namespace Microsoft.Azure.Cosmos
         /// <summary>
         /// Gets the size of documents outputted in bytes during query in the Azure DocumentDB database service.
         /// </summary>
-        internal long OutputDocumentSize
+        public long OutputDocumentSize
         {
             get
             {
@@ -168,7 +168,7 @@ namespace Microsoft.Azure.Cosmos
         /// <summary>
         /// Gets the total query time in the Azure DocumentDB database service.
         /// </summary>
-        internal TimeSpan TotalQueryExecutionTime
+        public TimeSpan TotalQueryExecutionTime
         {
             get
             {
@@ -212,7 +212,7 @@ namespace Microsoft.Azure.Cosmos
         /// <summary>
         /// Gets the query index lookup time in the Azure DocumentDB database service.
         /// </summary>
-        internal TimeSpan IndexLookupTime
+        public TimeSpan IndexLookupTime
         {
             get
             {
@@ -223,7 +223,7 @@ namespace Microsoft.Azure.Cosmos
         /// <summary>
         /// Gets the document loading time during query in the Azure DocumentDB database service.
         /// </summary>
-        internal TimeSpan DocumentLoadTime
+        public TimeSpan DocumentLoadTime
         {
             get
             {
@@ -234,7 +234,7 @@ namespace Microsoft.Azure.Cosmos
         /// <summary>
         /// Gets the query runtime execution times during query in the Azure DocumentDB database service.
         /// </summary>
-        internal RuntimeExecutionTimes RuntimeExecutionTimes
+        public RuntimeExecutionTimes RuntimeExecutionTimes
         {
             get
             {
@@ -245,7 +245,7 @@ namespace Microsoft.Azure.Cosmos
         /// <summary>
         /// Gets the output writing/serializing time during query in the Azure DocumentDB database service.
         /// </summary>
-        internal TimeSpan DocumentWriteTime
+        public TimeSpan DocumentWriteTime
         {
             get
             {
@@ -281,7 +281,7 @@ namespace Microsoft.Azure.Cosmos
         /// <summary>
         /// Gets the Index Hit Document Count.
         /// </summary>
-        internal long IndexHitDocumentCount
+        public long IndexHitDocumentCount
         {
             get
             {
@@ -292,7 +292,7 @@ namespace Microsoft.Azure.Cosmos
         /// <summary>
         /// Gets the VMExecution Time.
         /// </summary>
-        internal TimeSpan VMExecutionTime
+        public TimeSpan VMExecutionTime
         {
             get
             {
@@ -560,7 +560,7 @@ namespace Microsoft.Azure.Cosmos
         /// <summary>
         /// Gets the query runtime execution times during query in the Azure DocumentDB database service.
         /// </summary>
-        internal RuntimeExecutionTimes RuntimeExecutionTimes
+        public RuntimeExecutionTimes RuntimeExecutionTimes
         {
             get
             {